# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.APIError'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  services.GrantScopedAdminRequest:
    properties:
      admin_level:
        type: integer
      notes:
        type: string
      operator_user_id:
        type: integer
      operator_username:
        type: string
      permissions:
        items:
          type: string
        type: array
      project_ids:
        items:
          type: string
        type: array
      scope_type:
        description: '"global" or "projects"'
        type: string
      system_role_id:
        type: integer
      target_user_id:
        type: integer
    type: object
  services.GrantSystemAdminRequest:
    properties:
      admin_level:
        type: integer
      admin_scopes:
        additionalProperties: true
        type: object
      notes:
        type: string
      operator_user_id:
        type: integer
      operator_username:
        type: string
      permissions:
        items:
          type: string
        type: array
      project_ids:
        items:
          type: string
        type: array
      scope_type:
        description: '"global" or "scoped"'
        type: string
      system_role_id:
        type: integer
      target_user_id:
        type: integer
    type: object
...
  services.RevokeSystemAdminRequest:
    properties:
      operator_user_id:
        type: integer
      operator_username:
        type: string
      reason:
        type: string
      target_user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI项目管理后端API服务，提供项目管理、任务管理、文档生成等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Project Backend API
  version: "1.0"
paths:
  /admin/tasks/{taskId}/project:
    put:
      consumes:
      - application/json
      description: Change the project that a task belongs to (System administrators
        only)
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Update task project request
        in: body
        name: request
        required: true
        schema:
          properties:
            new_project_id:
              type: integer
            reason:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Task project updated successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden - System admin only
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Task or project not found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update task project
      tags:
      - system-admin
      - tasks
...
  /api/v1/system/admins:
    get:
      description: Get list of all active system administrators
      parameters:
      - description: Minimum admin level
        in: query
        name: min_level
        type: integer
      - description: Maximum admin level
        in: query
        name: max_level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List system admins
      tags:
      - system-admin
  /api/v1/system/admins/{id}/accessible-projects:
    get:
      description: Get list of projects that an admin can access
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get accessible projects
      tags:
      - system-admin
  /api/v1/system/admins/audit-logs:
    get:
      description: Get audit logs for system admin operations
      parameters:
      - description: Filter by target user ID
        in: query
        name: target_user_id
        type: integer
      - description: Filter by operator user ID
        in: query
        name: operator_user_id
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get audit logs
      tags:
      - system-admin
  /api/v1/system/admins/check:
    get:
      description: Check if the current authenticated user is a system administrator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Check current user
      tags:
      - system-admin
  /api/v1/system/admins/grant:
    post:
      consumes:
      - application/json
      description: Grant system administrator privileges to a user
      parameters:
      - description: Grant request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.GrantSystemAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Grant system admin
      tags:
      - system-admin
  /api/v1/system/admins/grant-scoped:
    post:
      consumes:
      - application/json
      description: Grant system admin privileges with specific project scope
      parameters:
      - description: Grant scoped request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.GrantScopedAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Grant scoped admin
      tags:
      - system-admin
  /api/v1/system/admins/revoke:
    post:
      consumes:
      - application/json
      description: Revoke system administrator privileges from a user
      parameters:
      - description: Revoke request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.RevokeSystemAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Revoke system admin
      tags:
      - system-admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"